Add files via upload
Exploratory data analysis on the all_speeches.csv and the sentiment words, which includes, data cleaning, readjusting the data with the  Edward George served as Governor of the Bank of England from 1 July 1993 to 30 June 2003. Date cleaning and outlier detection, heatmap and pairplot to visualise the correlations , violin plot and boxplot to visualise outliers. Also carried out Skewness and Kurtosis to check for skewness and distribution, used Yeo-Johnson transformation for normalising the data distribution. Outliers exist but decided to ignore as they are common in financial data.